store/backend: move host detection helpers to composerApi
The getHostArch and getHostDistro helpers were scattered across the
codebase. Centralise them under the on-prem composerApi helpers
directory so they can back a new RTK Query endpoint.
