+++ src/store/api/backend/onprem/composerApi/helpers/index.ts
@@ -3,6 +3,7 @@ export { parseJsonUnsafe } from './parseJson';
 export { lookupDatastreamDistro } from './dataStreamLookup';
 export { getBlueprintsPath } from './getBlueprintsPath';
 export { getCloudConfigs } from './getCloudConfigs';
+export { getHostArch, getHostDistro } from './hostInfo';
 export {
   byDistroDescending,
   inferDistro,
